pyproject.toml
[project]
name = "docscan-proto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "ipython>=9.7.0",
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "pylsd-nova>=1.2.1",
    "pyzbar>=0.1.9",
    "scipy>=1.16.3",
    "zxing-cpp>=2.3.0",
]
